use action container
When external workflows reference this action, they always hit
`/action.yml`.  Since that referrs to a Dockerfile, an image is always
re-built, which is time consuming and inefficient.  However, when
testing pull-requests against this repository, it's desireable to
have them utilize the Dockerfile.  This allows testing of changes
to the Dockerfile itself.

Support both cases by in-line modifying the `actions.yml` file to utilize
the Dockerfile when testing PRs.  Otherwise, this file should referr
to the image and tag produced by a previous run of the 'push' workflow.
This enables continuous delivery of the action and a pre-built
container.

Signed-off-by: Chris Evich <cevich@redhat.com>